backend/requirements.txt

fastapi
uvicorn
python-multipart
python-dotenv
google-generativeai
pillow
requests
